# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MovieInfo API",
    "description": "API documentation for MovieInfo application",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/watchlists/update/{id}": {
      "put": {
        "tags": [
          "watchlist-controller"
        ],
        "operationId": "updateWatchlist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/update/{id}": {
      "put": {
        "tags": [
          "user-controller"
        ],
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/update/{id}": {
      "put": {
        "tags": [
          "review-controller"
        ],
        "operationId": "updateReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/movies/update/{id}": {
      "put": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "updateMovie",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MovieResponse"
                }
              }
            }
          }
        }
      }
    },
    "/watchlists": {
      "post": {
        "tags": [
          "watchlist-controller"
        ],
        "operationId": "createWatchlist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResponse"
                }
              }
            }
          }
        }
      }
    },
    "/watched-movies/mark-as-watched": {
      "post": {
        "tags": [
          "watched-movie-controller"
        ],
        "operationId": "markAsWatched",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchedMovieRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/add": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/add": {
      "post": {
        "tags": [
          "review-controller"
        ],
        "operationId": "addReview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/movies/add": {
      "post": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "addMovie",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MovieResponse"
                }
              }
            }
          }
        }
      }
    },
    "/watchlists/get/{id}": {
      "get": {
        "tags": [
          "watchlist-controller"
        ],
        "operationId": "getWatchlistById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/movies/sorted-by-year": {
      "get": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "getAllMoviesSortedByReleaseYear",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MovieResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/search": {
      "get": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "searchMovies",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "director",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MovieResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/get_details/{userId}": {
      "get": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "getMovies",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MovieUserDetails"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/all": {
      "get": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "getAllMovies",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MovieResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/watchlists/delete/{id}": {
      "delete": {
        "tags": [
          "watchlist-controller"
        ],
        "operationId": "deleteWatchlist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/delete/{id}": {
      "delete": {
        "tags": [
          "user-controller"
        ],
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/reviews/delete/{id}": {
      "delete": {
        "tags": [
          "review-controller"
        ],
        "operationId": "deleteReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/movies/delete/{id}": {
      "delete": {
        "tags": [
          "movie-controller"
        ],
        "operationId": "deleteMovie",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "WatchlistRequest": {
        "required": [
          "name",
          "userId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "movieIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "WatchlistResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "movieTitles": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UserRequest": {
        "required": [
          "email",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "ReviewRequest": {
        "required": [
          "movieId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "movieId": {
            "type": "integer",
            "format": "int64"
          },
          "rating": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "maxLength": 5000,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ReviewResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "comment": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "movieId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "MovieRequest": {
        "required": [
          "actors",
          "director",
          "genre",
          "releaseYear",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 250,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 5000,
            "minLength": 0,
            "type": "string"
          },
          "releaseYear": {
            "maximum": 2030,
            "minimum": 1880,
            "type": "integer",
            "format": "int32"
          },
          "genre": {
            "type": "string",
            "enum": [
              "ACTION",
              "COMEDY",
              "DRAMA",
              "HORROR",
              "SCI_FI",
              "DOCUMENTARY",
              "ANIMATION",
              "THRILLER",
              "ROMANCE",
              "ADVENTURE"
            ]
          },
          "director": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "actors": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "MovieResponse": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "releaseYear": {
            "type": "integer",
            "format": "int32"
          },
          "genre": {
            "type": "string"
          },
          "directorName": {
            "type": "string"
          },
          "actors": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "WatchedMovieRequest": {
        "required": [
          "movieId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "movieId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "MovieUserDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "releaseYear": {
            "type": "integer",
            "format": "int32"
          },
          "genre": {
            "type": "string"
          },
          "directorName": {
            "type": "string"
          },
          "watchlists": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reviews": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "actors": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "watched": {
            "type": "boolean"
          }
        }
      }
    }
  }
}