# src/main/java/com/example/movieInfo/dto/MovieResponse.java
package com.example.movieInfo.dto;

import java.util.Set;

public class MovieResponse {
    private String title;
    private String description;
    private Integer releaseYear;
    private String genre;
    private String directorName;
    private Set<String> actors;

    public MovieResponse(){
    }

    public MovieResponse(String title, String description, Integer releaseYear, String genre, String directorName, Set<String> actors) {
        this.title = title;
        this.description = description;
        this.releaseYear = releaseYear;
        this.genre = genre;
        this.directorName = directorName;
        this.actors = actors;
    }


    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public Integer getReleaseYear() {
        return releaseYear;
    }

    public void setReleaseYear(Integer releaseYear) {
        this.releaseYear = releaseYear;
    }

    public String getGenre() {
        return genre;
    }

    public void setGenre(String genre) {
        this.genre = genre;
    }

    public String getDirectorName() {
        return directorName;
    }

    public void setDirectorName(String directorName) {
        this.directorName = directorName;
    }

    public Set<String> getActors() {
        return actors;
    }

    public void setActors(Set<String> actors) {
        this.actors = actors;
    }
}
