# src/main/java/com/example/movieInfo/dto/ReviewResponse.java
package com.example.movieInfo.dto;

public class ReviewResponse {
    private Long id;
    private Double rating;
    private String comment;
    private Long userId;
    private Long movieId;

    public ReviewResponse(){

    }

    public ReviewResponse(Long id, Double rating, String comment, Long userId, Long movieId) {
        this.id = id;
        this.rating = rating;
        this.comment = comment;
        this.userId = userId;
        this.movieId = movieId;
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Double getRating() {
        return rating;
    }

    public void setRating(Double rating) {
        this.rating = rating;
    }

    public String getComment() {
        return comment;
    }

    public void setComment(String comment) {
        this.comment = comment;
    }

    public Long getUserId() {
        return userId;
    }

    public void setUserId(Long userId) {
        this.userId = userId;
    }

    public Long getMovieId() {
        return movieId;
    }

    public void setMovieId(Long movieId) {
        this.movieId = movieId;
    }
}
